This dummy copied the packages:index file and worked on the wrong copy with the truncate command. Put the right commands in the right files.
